// src/app/shared/models/location.ts
// Generated by https://quicktype.io
export interface ILocation {
    longitude: number;
    city: string;
    timezone: string;
    offset: number;
    region: string;
    asn: number;
    organization: string;
    country: string;
    ip: string;
    latitude: number;
    postal_code: string;
    continent_code: string;
    country_code: string;
    region_code: string;
}